Feat: feign client added

+++ storehouse-service/src/main/java/org/dei/storehouse/controller/StoreHouseController.java
@@ -14,18 +14,18 @@ public interface StoreHouseController {
     List<Product> getItems();
 
     @RequestMapping(path = "/items/{item_id}", method = RequestMethod.GET)
-    Product getItemById(@PathVariable(value = "item_id") int itemID);
+    Product getItemById(@PathVariable(value = "item_id") Long itemID);
 
     @RequestMapping(path = "/items", method = RequestMethod.POST)
     Product createItem(@RequestBody ProductCreationDTO productCreationDTO);
 
     @RequestMapping(path = "/items/{id}/addition/{amount}", method = RequestMethod.PUT)
-    Product addItems(@PathVariable int id, @PathVariable Long amount);
+    Product addItems(@PathVariable Long id, @PathVariable Long amount);
 
     @RequestMapping(path = "/items/{id}/reserve/{amount}", method = RequestMethod.POST)
-    Product reserveItems(@PathVariable int id, @PathVariable Long amount);
+    Product reserveItems(@PathVariable Long id, @PathVariable Long amount);
 
     @RequestMapping(path = "/items/{id}/unreserve/{amount}", method = RequestMethod.POST)
-    Product unreserveItems(@PathVariable int id, @PathVariable Long amount);
+    Product unreserveItems(@PathVariable Long id, @PathVariable Long amount);
 
 }
